Nuke `Table` class
