+++ src/main/kotlin/sigma/evaluation/values/BoolValue.kt
@@ -11,7 +11,7 @@ data class BoolValue(
 
     object If : ComputableFunctionValue() {
         override fun apply(argument: Value): Value {
-            val test = (argument as Table).read(IntValue.Zero)!! as BoolValue
+            val test = (argument as DictValue).read(IntValue.Zero)!! as BoolValue
 
             return object : ComputableFunctionValue() {
                 override fun apply(argument: Value): Value {
